Move JournalOp to commonMain in new package
