+++ lrucache/src/jvmMain/kotlin/com/mayakapps/lrucache/Journal.kt
@@ -1,5 +1,6 @@
 package com.mayakapps.lrucache
 
+import com.mayakapps.lrucache.journal.JournalOp
 import java.io.*
 
 // This is more than enough as the buffer is flushing after each operation
@@ -84,34 +85,5 @@ internal class JournalReader(journalFile: File) : Closeable {
 
     override fun close() {
         inputStream.close()
-    }
-}
-
-internal sealed interface JournalOp {
-    val key: String
-
-    data class Dirty(override val key: String) : JournalOp
-    data class Clean(override val key: String) : JournalOp
-    data class Remove(override val key: String) : JournalOp
-
-    val opcode: Byte
-        get() = when (this) {
-            is Dirty -> OPCODE_DIRTY
-            is Clean -> OPCODE_CLEAN
-            is Remove -> OPCODE_REMOVE
-        }
-
-    companion object {
-        fun create(opcode: Byte, key: String) =
-            when (opcode) {
-                OPCODE_DIRTY -> Dirty(key)
-                OPCODE_CLEAN -> Clean(key)
-                OPCODE_REMOVE -> Remove(key)
-                else -> throw IllegalArgumentException()
-            }
-
-        private const val OPCODE_DIRTY: Byte = 1
-        private const val OPCODE_CLEAN: Byte = 2
-        private const val OPCODE_REMOVE: Byte = 3
     }
 }
